requirements.txt
# Framework
fastapi==0.70.0

# aws
boto3==1.19.4
pynamodb==5.1.0

# ASGI Server
uvicorn[standard]==0.15.0

# Dependencies
bcrypt==3.2.0
elasticsearch==7.15.1
numpy==1.21.3
pandas==1.3.4
passlib==1.7.4
pydantic==1.8.2
python-multipart==0.0.5
PyJWT==2.3.0
scikit-learn==1.0.1
scipy==1.7.1
sentry-sdk==1.4.3
sqlalchemy==1.4.26
requests==2.26.0
mysql-connector-python==8.0.27

# Middleware
ddtrace==0.55.0

# For CLI
typer==0.4.0
